refactor(reports): Moved models to reports
