+++ reports/tests/test_put_report.py
@@ -1,9 +1,9 @@
 import json
 from unittest import mock
 
-from db_handler.models import Report
 from pymongo.errors import DuplicateKeyError, ServerSelectionTimeoutError
 
+from reports.models import Report
 from . import utils
 
 oid = utils.random_oid()
